+++ src/features/updates/components/ChapterUpdateCard.tsx
@@ -20,18 +20,12 @@ import { ChapterDownloadButton } from '@/features/chapter/components/buttons/Cha
 import { ChapterDownloadRetryButton } from '@/features/chapter/components/buttons/ChapterDownloadRetryButton.tsx';
 import { Chapters } from '@/features/chapter/services/Chapters.ts';
 import { ListCardContent } from '@/base/components/lists/cards/ListCardContent.tsx';
-import { plural } from '@lingui/core/macro';
 import Stack from '@mui/material/Stack';
-import Button from '@mui/material/Button';
-import { MUIUtil } from '@/lib/mui/MUI.util.ts';
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import ExpandLessIcon from '@mui/icons-material/ExpandLess';
 import Collapse from '@mui/material/Collapse';
 import { STABLE_EMPTY_ARRAY } from '@/base/Base.constants.ts';
-import { useTheme } from '@mui/material/styles';
-import { TypographyMaxLines } from '@/base/components/texts/TypographyMaxLines.tsx';
 import { Virtuoso } from 'react-virtuoso';
 import { ReaderService } from '@/features/reader/services/ReaderService.ts';
+import { ChapterCardExpandButton } from '@/features/chapter/components/buttons/ChapterCardExpandButton.tsx';
 
 export const ChapterUpdateCard = memo(
     ({
@@ -42,8 +36,6 @@ export const ChapterUpdateCard = memo(
         otherChapters?: ChapterUpdateListFieldsFragment[];
     }) => {
         const { manga } = chapter;
-
-        const theme = useTheme();
 
         const [isExpanded, setIsExpanded] = useState(false);
 
@@ -61,7 +53,7 @@ export const ChapterUpdateCard = memo(
                         ReaderService.openReader(chapter);
                     }}
                     sx={{
-                        color: theme.palette.text[chapter.isRead ? 'disabled' : 'primary'],
+                        color: (theme) => theme.palette.text[chapter.isRead ? 'disabled' : 'primary'],
                     }}
                 >
                     <ListCardContent sx={{ justifyContent: 'space-between' }}>
@@ -105,29 +97,11 @@ export const ChapterUpdateCard = memo(
                                     ternaryText={chapter.name}
                                 />
                                 {isGroup && (
-                                    <Button
-                                        sx={{
-                                            maxWidth: 'fit-content',
-                                            justifyContent: 'flex-start',
-                                            ...theme.typography.caption,
-                                        }}
-                                        variant="text"
-                                        size="small"
-                                        endIcon={isExpanded ? <ExpandLessIcon /> : <ExpandMoreIcon />}
-                                        {...MUIUtil.preventRippleProp()}
-                                        onClick={(e) => {
-                                            e.preventDefault();
-                                            e.stopPropagation();
-                                            setIsExpanded(!isExpanded);
-                                        }}
-                                    >
-                                        <TypographyMaxLines lines={1} variant="caption" sx={{ textAlign: 'start' }}>
-                                            {plural(otherChapters.length, {
-                                                one: 'Show # more chapter',
-                                                other: 'Show # more chapters',
-                                            })}
-                                        </TypographyMaxLines>
-                                    </Button>
+                                    <ChapterCardExpandButton
+                                        count={otherChapters.length}
+                                        expanded={isExpanded}
+                                        setExpanded={setIsExpanded}
+                                    />
                                 )}
                             </Stack>
                         </Box>